differentiate between provided classes and those to change via packages
+ some cleanup
