--- src/main/java/com/ververica/flinktraining/exercises/troubleshoot/source/ObjectMapperSingleton.java
@@ -1,13 +1,0 @@
-package com.ververica.flinktraining.exercises.troubleshoot.source;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class ObjectMapperSingleton {
-
-    public static ObjectMapper getInstance() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        return objectMapper;
-    }
-}
