+++ src/main/java/com/ververica/flinktraining/provided/troubleshoot/FakeKafkaRecord.java
@@ -1,8 +1,11 @@
-package com.ververica.flinktraining.exercises.troubleshoot.entities;
+package com.ververica.flinktraining.provided.troubleshoot;
+
+import com.ververica.flinktraining.provided.DoNotChangeThis;
 
 import java.util.Arrays;
 import java.util.Objects;
 
+@DoNotChangeThis
 public class FakeKafkaRecord {
 
     private long   timestamp;
