+++ src/main/java/com/ververica/flinktraining/solutions/troubleshoot/SimpleMeasurement.java
@@ -1,30 +1,20 @@
-package com.ververica.flinktraining.exercises.troubleshoot.entities;
+package com.ververica.flinktraining.solutions.troubleshoot;
 
 import java.util.Objects;
 
-public class Measurement {
+public class SimpleMeasurement {
 
     private int    sensorId;
     private double value;
     private String location;
-    private String measurementInformation;
 
-    public Measurement() {
+    public SimpleMeasurement() {
     }
 
-    public Measurement(final int sensorId, final double value, final String location, final String measurementInformation) {
+    public SimpleMeasurement(final int sensorId, final double value, final String location, final String measurementInformation) {
         this.sensorId = sensorId;
         this.value = value;
         this.location = location;
-        this.measurementInformation = measurementInformation;
-    }
-
-    public String getMeasurementInformation() {
-        return measurementInformation;
-    }
-
-    public void setMeasurementInformation(final String measurementInformation) {
-        this.measurementInformation = measurementInformation;
     }
 
     public int getSensorId() {
@@ -59,16 +49,16 @@ public class Measurement {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        final Measurement that = (Measurement) o;
+        final SimpleMeasurement
+                that = (SimpleMeasurement) o;
         return sensorId == that.sensorId &&
                Double.compare(that.value, value) == 0 &&
-               Objects.equals(location, that.location) &&
-               Objects.equals(measurementInformation, that.measurementInformation);
+               Objects.equals(location, that.location);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(sensorId, value, location, measurementInformation);
+        return Objects.hash(sensorId, value, location);
     }
 
     @Override
@@ -77,7 +67,6 @@ public class Measurement {
         sb.append("sensorId=").append(sensorId);
         sb.append(", value=").append(value);
         sb.append(", location='").append(location).append('\'');
-        sb.append(", measurementInformation='").append(measurementInformation).append('\'');
         sb.append('}');
         return sb.toString();
     }
