+++ src/main/java/com/ververica/flinktraining/solutions/troubleshoot/TroubledStreamingJobSolution41.java
@@ -19,11 +19,12 @@ import org.apache.flink.streaming.api.windowing.windows.TimeWindow;
 import org.apache.flink.util.Collector;
 import org.apache.flink.util.OutputTag;
 
-import com.ververica.flinktraining.exercises.troubleshoot.entities.FakeKafkaRecord;
-import com.ververica.flinktraining.exercises.troubleshoot.entities.Measurement;
-import com.ververica.flinktraining.exercises.troubleshoot.entities.WindowedMeasurements;
-import com.ververica.flinktraining.exercises.troubleshoot.source.ObjectMapperSingleton;
-import com.ververica.flinktraining.exercises.troubleshoot.source.SourceUtils;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ververica.flinktraining.provided.troubleshoot.FakeKafkaRecord;
+import com.ververica.flinktraining.provided.troubleshoot.Measurement;
+import com.ververica.flinktraining.provided.troubleshoot.WindowedMeasurements;
+import com.ververica.flinktraining.provided.troubleshoot.SourceUtils;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
@@ -138,11 +139,7 @@ public class TroubledStreamingJobSolution41 {
         private final long maxOutOfOrderness;
         private final long idleTimeout;
 
-        public MeasurementTSExtractor() {
-            this(Time.of(250, TimeUnit.MILLISECONDS), Time.of(1, TimeUnit.SECONDS));
-        }
-
-        public MeasurementTSExtractor(Time maxOutOfOrderness, Time idleTimeout) {
+        MeasurementTSExtractor(Time maxOutOfOrderness, Time idleTimeout) {
             if (maxOutOfOrderness.toMilliseconds() < 0) {
                 throw new RuntimeException("Tried to set the maximum allowed " +
                         "lateness to " + maxOutOfOrderness +
@@ -232,7 +229,7 @@ public class TroubledStreamingJobSolution41 {
 
         private transient DescriptiveStatisticsHistogram eventTimeLag;
 
-        public MeasurementWindowProcessFunction() {
+        MeasurementWindowProcessFunction() {
         }
 
         @Override
@@ -260,6 +257,14 @@ public class TroubledStreamingJobSolution41 {
 
             eventTimeLag = getRuntimeContext().getMetricGroup().histogram("eventTimeLag",
                     new DescriptiveStatisticsHistogram(EVENT_TIME_LAG_WINDOW_SIZE));
+        }
+    }
+
+    private static class ObjectMapperSingleton {
+        static ObjectMapper getInstance() {
+            ObjectMapper objectMapper = new ObjectMapper();
+            objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+            return objectMapper;
         }
     }
 }
