+++ src/main/java/com/ververica/flinktraining/solutions/troubleshoot/TroubledStreamingJobSolution43.java
@@ -19,12 +19,11 @@ import org.apache.flink.streaming.api.windowing.windows.TimeWindow;
 import org.apache.flink.util.Collector;
 import org.apache.flink.util.OutputTag;
 
+import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ververica.flinktraining.exercises.troubleshoot.entities.FakeKafkaRecord;
-import com.ververica.flinktraining.exercises.troubleshoot.entities.SimpleMeasurement;
-import com.ververica.flinktraining.exercises.troubleshoot.entities.WindowedMeasurements;
-import com.ververica.flinktraining.exercises.troubleshoot.source.ObjectMapperSingleton;
-import com.ververica.flinktraining.exercises.troubleshoot.source.SourceUtils;
+import com.ververica.flinktraining.provided.troubleshoot.FakeKafkaRecord;
+import com.ververica.flinktraining.provided.troubleshoot.WindowedMeasurements;
+import com.ververica.flinktraining.provided.troubleshoot.SourceUtils;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
@@ -70,7 +69,8 @@ public class TroubledStreamingJobSolution43 {
                 .keyBy(SimpleMeasurement::getLocation)
                 .timeWindow(Time.of(1, TimeUnit.SECONDS))
                 .sideOutputLateData(lateDataTag)
-                .aggregate(new MeasurementWindowAggregatingFunction(), new MeasurementWindowProcessFunction())
+                .aggregate(new MeasurementWindowAggregatingFunction(),
+                        new MeasurementWindowProcessFunction())
                 .name("WindowedAggregationPerLocation")
                 .uid("WindowedAggregationPerLocation");
 
@@ -101,7 +101,7 @@ public class TroubledStreamingJobSolution43 {
      * Deserializes the JSON Kafka message.
      */
     public static class MeasurementDeserializer extends RichFlatMapFunction<FakeKafkaRecord, SimpleMeasurement> {
-        private static final long serialVersionUID = 4054149949298485680L;
+        private static final long serialVersionUID = 4L;
 
         private Counter numInvalidRecords;
         private transient ObjectMapper instance;
@@ -110,7 +110,7 @@ public class TroubledStreamingJobSolution43 {
         public void open(final Configuration parameters) throws Exception {
             super.open(parameters);
             numInvalidRecords = getRuntimeContext().getMetricGroup().counter("numInvalidRecords");
-            instance = ObjectMapperSingleton.getInstance();
+            instance = createObjectMapper();
         }
 
         @Override
@@ -139,10 +139,6 @@ public class TroubledStreamingJobSolution43 {
 
         private final long maxOutOfOrderness;
         private final long idleTimeout;
-
-        public MeasurementTSExtractor() {
-            this(Time.of(250, TimeUnit.MILLISECONDS), Time.of(1, TimeUnit.SECONDS));
-        }
 
         MeasurementTSExtractor(Time maxOutOfOrderness, Time idleTimeout) {
             if (maxOutOfOrderness.toMilliseconds() < 0) {
@@ -263,5 +259,11 @@ public class TroubledStreamingJobSolution43 {
             eventTimeLag = getRuntimeContext().getMetricGroup().histogram("eventTimeLag",
                     new DescriptiveStatisticsHistogram(EVENT_TIME_LAG_WINDOW_SIZE));
         }
+    }
+
+    private static ObjectMapper createObjectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        return objectMapper;
     }
 }
